move and rename beacon_state_types to eth_spec.rs
