+++ eth2/types/src/beacon_state.rs
@@ -15,10 +15,9 @@ use tree_hash::TreeHash;
 use tree_hash_derive::TreeHash;
 
 pub use self::committee_cache::CommitteeCache;
-pub use beacon_state_types::*;
+pub use eth_spec::*;
 
 #[macro_use]
-mod beacon_state_types;
 mod committee_cache;
 mod exit_cache;
 mod pubkey_cache;
