+++ eth2/types/src/lib.rs
@@ -23,6 +23,7 @@ pub mod crosslink_committee;
 pub mod deposit;
 pub mod deposit_data;
 pub mod eth1_data;
+pub mod eth_spec;
 pub mod fork;
 pub mod free_attestation;
 pub mod historical_batch;
